# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-reciprocal",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `1 / value` by delegating to `srvcs-floatdivide`.",
        "description": "This service owns the *control flow* but delegates the arithmetic to its\ndependency, exactly as specified:\n\n1. ask `srvcs-floatdivide` for `1 / value`;\n2. surface its `result` as the reciprocal.\n\nIf `value` is `0`, floatdivide rejects the divisor with `422`, which is\nforwarded verbatim. If the dependency is unreachable it reports itself\ndegraded (`503`).",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReciprocalResponse"
                }
              }
            }
          },
          "422": {
            "description": "the dependency rejected the input (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "ReciprocalResponse": {
        "type": "object",
        "required": [
          "value",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double",
            "description": "The reciprocal `1 / value` as a floating-point number."
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}